Convert chatbox to razor
