+++ code/UI/Styles/chatbox.scss
@@ -1,6 +1,7 @@
-﻿@import "../Vars.scss";
+﻿
+@import "_vars.scss";
 
-UfChatbox {
+UfChat {
   position: absolute;
   bottom: 64px;
   left: 64px;
@@ -14,7 +15,7 @@ UfChatbox {
   font-size: 20px;
 
   &.topleft {
-      top: 64px;
+    top: 64px;
   }
 
   &.open {
@@ -57,7 +58,7 @@ UfChatbox {
   }
 }
 
-UfChatboxEntry {
+UfChatEntry {
   padding-top: 8px;
   flex-shrink: 0;
   flex-grow: 0;
@@ -113,7 +114,7 @@ UfChatboxEntry {
     }
 
     .message {
-        background-color: $orange-light;
+      background-color: $orange-light;
     }
 
     &.timer-msg {
