+++ code/UI/Styles/timerhud.scss
@@ -1,4 +1,4 @@
-﻿@import "Vars.scss";
+﻿@import "_vars.scss";
 
 TimerHud {
   justify-content: center;
