+++ code/UI/TimerHud.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StyleSheet( "/UI/TimerHud.scss" )]
+﻿@attribute [StyleSheet( "/UI/Styles/stylesheet.scss" )]
 
 <root class=@RootClass>
     <div class="game-timer">
