--- code/UI/UfChatbox/UfChatbox.html
@@ -1,12 +1,0 @@
-﻿
-<link rel="stylesheet" href="UfChatbox.scss" />
-
-<root>
-
-    <div @ref="EntryCanvas" class="entries"></div>
-
-    <div class="input">
-        <textentry @ref="Input" placeholder="hi"></textentry>
-    </div>
-
-</root>
