--- code/UI/UfChatbox/UfChatboxEntry.html
@@ -1,5 +1,0 @@
-﻿
-<root>
-    <label @ref="Name" class="name"></label>
-    <label @ref="Message" class="message"></label>
-</root>
