+++ code/UI/UnicycleHud.html
@@ -6,7 +6,7 @@
     <TimerHud />
     <NewPlayerOverlay></NewPlayerOverlay>
     <GameMenu></GameMenu>
-    <UfChatbox></UfChatbox>
+    <UfChat />
     <UfKillfeed></UfKillfeed>
     <FrenzyCollectionHud></FrenzyCollectionHud>
     <GameEnd></GameEnd>
