+++ code/UI/_Imports.cs
@@ -3,3 +3,5 @@ global using SandboxEditor;
 global using Sandbox.UI;
 global using System;
 global using System.Threading.Tasks;
+global using System.Collections;
+global using System.Collections.Generic;
