+++ code/UnicycleFrenzy.State.cs
@@ -20,13 +20,13 @@ internal partial class UnicycleFrenzy
 			await Task.DelayRealtimeSeconds( 1.0f );
 		}
 
-		UfChatbox.AddChat( To.Everyone, "Server", "The match will start in 10 seconds", "alert", "chat.alert" );
+		UfChat.AddChat( To.Everyone, "Server", "The match will start in 10 seconds", "alert", "chat.alert" );
 
 		GameState = GameStates.PreMatch;
 		StateTimer = 5f;
 		await WaitStateTimer();
 
-		UfChatbox.AddChat( To.Everyone, "Server", "The match is live!", "alert", "chat.alert" );
+		UfChat.AddChat( To.Everyone, "Server", "The match is live!", "alert", "chat.alert" );
 
 		GameState = GameStates.Live;
 		StateTimer = 15f * 60;
@@ -91,7 +91,7 @@ internal partial class UnicycleFrenzy
 		var needed = 3;
 		var current = Client.All.Count;
 
-		UfChatbox.AddChat( To.Everyone, "Server", $"{current} out of {needed} players needed to start", "alert", "chat.alert" );
+		UfChat.AddChat( To.Everyone, "Server", $"{current} out of {needed} players needed to start", "alert", "chat.alert" );
 	}
 
 	[ConCmd.Admin]
